ensure calculation ends up with a positive integer
I've been using the utils to find the centre and zoom needed to show a number of markers on the visible map. Under some conditions we get no zoom level. I believe it's because we have some markers with negative latitude and some with positive. An Math.abs fixes the problem.