Documentation: added PropTypes for heatmap prop (#716)
* Documentation: added PropTypes for heatmap prop

This PR improves the documentation of the heatmap prop by adding types.
I don't mind adding PropTypes validation to this prop as well if you are interested in that.

* Update API.md

I added some spacing between the interface and the example snippet.
I also added a link to the example I wrote for the other repository. I think we should delay this PR until that example is live. That way, the new Demo link will go to live functioning code rather than no where.