add example linking to react-table search
I really love google-map-react, and thought I'd share a link to my project that uses it and react-table, to help comedians find comedy nights to perform at. 

The interesting thing I've done is link the elements in the table to those shown on the map. So you can search in the table and the map updates.